+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Change Log
+
+## 1.0.17 - December 15, 2020
+
+### Changes
+
+- added `useBreadcrumb` hook to allow getting a path to the current page
+- updated `nav.js` component to accept links instead of filtering default menu links, allows for custom use cases.
+- refactor `SubNavigation` (`nav-submenu.js`) to use `Navigation` (`nav.js`) component